requirements.txt
coloredlogs
jadi
pyyaml
mako
python-augeas